WIP: Remove uses of deprecated JSON API
The version 3.12 of the Java driver deprecates `com.mongodb.util.JSON`
in favour of the JSON functions on the DBObject class. The deprecated
API will be removed in driver version 4.x.

In order to prepare for the upgrade to 4.x we can upgrade from driver
`3.10` to `3.12` and remove calls to the decprecated API.